+++ lib/features/shift/domain/usecases/close_shift_usecase.dart
@@ -8,7 +8,7 @@ class CloseShiftUseCase {
   final ShiftRepository repository;
   CloseShiftUseCase(this.repository);
 
-  Future<Either<Failure, Shift>> execute({required int shiftId, required double actualCash}) async {
+  Future<Either<Failure, Shift>> execute({required int shiftId, required int actualCash}) async {
     if (actualCash < 0) return const Left(CacheFailure('النقدية الفعلية لا يمكن أن تكون بالسالب'));
     return await repository.closeShift(shiftId: shiftId, actualCash: actualCash);
   }
